# Add performance tests for PBKDF2