--- a/browserTest/performance.html
+++ b/browserTest/performance.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>SJCL browser performance</title>
+  <link rel="stylesheet" type="text/css" href="test.css"/>
+</head>
+<body>
+  <h1>SJCL browser performance</h1>
+  <div id="status">Waiting for tests to begin...</div>
+  <div id="print"></div>
+  <script type="text/javascript" src="../sjcl.js"></script>
+  <script type="text/javascript" src="browserUtil.js"></script>
+  <script type="text/javascript" src="performance.js"></script>
+</body>
+</html>