This is how the page looks right now:

<img width="870" alt="captura de pantalla 2016-05-29 a las 12 04 26 a m" src="https://cloud.githubusercontent.com/assets/138426/15631155/f33caace-2530-11e6-985d-760741b20f31.png">

Results for **MacBook Pro 2,7 GHz Intel Core i5**:

| iter  | time |
| ------------- | ------------- |
| 1000 | 12.603 ms |
| 2000 | 16.013 ms |
| 4000 | 23.180 ms |
| 8000 | 47.197 ms |
| 16000 | 93.533 ms |
| 32000 | 191.752 ms |
| 48000 | 279.633 ms |
| 64000 | 371.350 ms |

Results for **iPhone SE**:

| iter  | time |
| ------------- | ------------- |
| 1000 | 14.697 ms |
| 2000 | 25.298 ms |
| 4000 | 45.127 ms |
| 8000 | 90.638 ms |
| 16000 | 173.513 ms |
| 32000 | 346.902 ms |
| 48000 | 523.743 ms |
| 64000 | 695.372 ms |

Results for **Samsung S4 Mini**:

| iter  | time |
| ------------- | ------------- |
| 1000 | 90.532 ms |
| 2000 | 97.463 ms |
| 4000 | 184.423 ms |
| 8000 | 361.310 ms |
| 16000 | 755.023 ms |
| 32000 | 1550.385 ms |
| 48000 | 2361.818 ms |
| 64000 | 3420.530 ms |

I am not claiming this is a representative sample, but my rough guess is that we can increment the iteration count to about 10,000 without degrading too much the average user experience. While this number is not going to work for every use case, it provides a reasonable amount of security as of 2016.

cc/ @Nilos  ref/ #289