@Nilos 2016-05-29
Looks good to me!
As soon as https://jsperf.com/ is back online, we could also add a test there.